Refactor OTLP trace ingestion (#2119)
* Refactor otlp traces content type handling

* Refactor extracting data from otlp ingestion

* Move common OTLP to a separate file

* Move some code to the main utils package

* Extract another helper function

* fixup
